Replaced "Right Single Quotation Marks" with generic apostrophe
Apparently the quotation mark is the recommended character to use as an apostrophe according to the Unicode standard, but since this is a Japanese game, it is treated as a full-width character, which spaces it out from the next letter. Plus, 90% of the apostrophes use the normal one so why would these be different?